Restructured sender module for ChannelSender, demo auto-encoding (but not typed at ChannelSender level!)
